// scripts/utils.d.ts
export declare const COLORS: {
    background: string;
    primary: string;
    secondary: string;
    accent: string;
    success: string;
    warning: string;
    error: string;
    textMain: string;
    textMuted: string;
    border: string;
};
export declare function wrapSVG(content: string, width: number, height: number, customStyles?: string): string;
export declare function writeSVG(filename: string, svg: string): void;
//# sourceMappingURL=utils.d.ts.map